requirements.txt

Flask~=2.3.2
Flask-RESTful>=0.3.6
flask-cors==3.0.8
Jinja2>=2.10.1
requests>=2.25.0
requests-aws4auth==1.1.0
python-dotenv~=1.0.0
spotipy~=2.23.0